LPS-124324 Extract add group and edit group to new action command class
