+++ modules/apps/site/site-admin-web/src/main/resources/META-INF/resources/edit_site.jsp
@@ -84,7 +84,7 @@ if (layoutSetPrototypeId > 0) {
 
 <liferay-ui:success key='<%= SiteAdminPortletKeys.SITE_SETTINGS + "requestProcessed" %>' message="site-was-added" />
 
-<portlet:actionURL name="editGroup" var="editGroupURL">
+<portlet:actionURL name="/site_admin/edit_group" var="editGroupURL">
 	<portlet:param name="mvcPath" value="/edit_site.jsp" />
 </portlet:actionURL>
